base for rebuild templates
