+++ src/app/types.ts
@@ -1,38 +1,3 @@
-// cryptocurrency name should be lowercase, because it is used to find cryptocurrency by its slug
-export enum SupportedCryptocurrencies {
-  Bitcoin = 'bitcoin',
-  Ethereum = 'ethereum',
-  Tether = 'tether',
-  BNB = 'bnb',
-  USDCoin = 'usd-coin',
-  Solana = 'solana',
-  XRP = 'xrp',
-  Terra = 'terra-luna',
-  Cardano = 'cardano',
-  Dogecoin = 'dogecoin',
-  Avalanche = 'avalanche',
-  TerraUSD = 'terrausd',
-  Polkadot = 'polkadot-new',
-  BinanceUSD = 'binance-usd',
-  ShibaInu = 'shiba-inu',
-  WrappedBitcoin = 'wrapped-bitcoin',
-  Polygon = 'polygon',
-  Litecoin = 'litecoin',
-  Cosmos = 'cosmos',
-  Chainlink = 'chainlink',
-  Uniswap = 'uniswap',
-  BitcoinCash = 'bitcoin-cash',
-  ApeCoin = 'apecoin-ape',
-  Stellar = 'stellar',
-  Monero = 'monero',
-  EthereumClassic = 'ethereum-classic',
-  Filecoin = 'filecoin',
-  ThetaNetwork = 'theta-network',
-  TheGraph = 'the-graph',
-  BitcoinSV = 'bitcoin-sv',
-  Amp = 'amp',
-}
-
 export enum AppEnv {
   Develop = 'develop',
   Prod = 'prod',
