+++ src/shared/index.ts
@@ -1,2 +1,3 @@
 export * from './strings'
 export * from './types'
+export * from './enums'
