+++ static/mailTemplates/cryptoAlert/cryptoAlertTemplates.ts
@@ -1,0 +1,1 @@
+export {}
