+++ static/mailTemplates/index.ts
@@ -1,0 +1,1 @@
+export { cryptoAlertTemplateBase } from './cryptoAlert/cryptoAlertTemplateBase'
